Move http_loader::Connector to a separate module
